+++ client/lib/data/models/ongoing_warrant_position.dart
@@ -12,7 +12,7 @@ class OngoingWarrantPosition implements IBaseEntity {
 
   final String isin;
   final double price;
-  final int numberOfShares;
+  final double numberOfShares;
   final DateTime goodUntil;
   final EnterOrExitPositionType type;
